Extract AppFeature+Effects.swift for side effects and async actions
